Add files via upload
Program to generate reads from background PWM and a list of signal PWMs using rejection sampling (generates random sequences and rejects sequences that do not match). There is no threshold, as the program uses a probabilistic approach (calculates probability of match and if random number between 0-1 is lower, takes the read as match).